# Improve NMI performance for multi_vram_buffer ops