For a single 3-byte horizontal operation this saved 14 cycles from `flush_vram_update2`

With two `multi_vram_buffer_vert` operations totaling 36 tiles and 10 `one_vram_buffer` this saved 44 cycles. This looks like it works out to ~12 cycles per multi and 2 cycles per single.