@cogwheel 2023-09-20
Hmm... I was under the impression that the function just changes the address of the vram buffer and that the existing APIs would still be used (though I should've updated that comment).

But yes, if a project has custom code to insert operations into the buffer, then this will be a breaking change.

I'll see if I can keep some of the performance gains without the format change.