@cogwheel 2023-09-21
I guess most of the gains were elsewhere; going back to the old format only cost 2 cycles for the multi_vram_buffer ops.